feat: Add form_edit_akun_user.blade.php for admin to edit user account
ADMIN PANEL : Management User Data -- Form Edit User #51

The code changes in this commit add a new file `form_edit_akun_user.blade.php` in the `admin/data_user` directory. This file is used by the admin to edit the account details of a user.

This commit follows the established conventions in the repository.