fix: Prevent users from deleting other users' kajian
This commit fixes a bug in the `KajianController` where users were able to delete kajian belonging to other users. The code now checks if the authenticated user is the owner of the kajian before allowing the deletion. If the user is not the owner, they are redirected to their profile page with an error message. This change improves the security of the application by preventing unauthorized deletion of kajian.

The commit message follows the established convention of using a prefix to indicate the type of change (`fix` for bug fixes) and provides a clear and concise description of the changes made.